pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
dynamic = ["version"]
name = "OGTFinder"
description = "OGTFinder is an optimal growth temperature prediction tool for prokaryotes using proteome-derived features"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
license = "MIT"
license-files = ["LICENSE"]

dependencies = [
    "pandas==2.2.2",
    "numpy==2.0.0",
    "scikit-learn==1.5.1",
    "biopython>=1.84", 
]

[tool.hatch.build]
exclude = ["Data"]

[project.scripts]
ogtfinder = "ogtfinder.ogtfinder:main"

[tool.hatch.version]
path = "src/ogtfinder/__init__.py"

[project.urls]
Homepage = "https://github.com/SC-Git1/OGTFinder"

